+++ ksyun/config.go
@@ -24,6 +24,7 @@ import (
 	"github.com/KscSDK/ksc-sdk-go/service/knad"
 	"github.com/KscSDK/ksc-sdk-go/service/krds"
 	"github.com/KscSDK/ksc-sdk-go/service/mongodb"
+	"github.com/KscSDK/ksc-sdk-go/service/pdns"
 	"github.com/KscSDK/ksc-sdk-go/service/rabbitmq"
 	"github.com/KscSDK/ksc-sdk-go/service/sks"
 	"github.com/KscSDK/ksc-sdk-go/service/slb"
@@ -48,6 +49,7 @@ type Config struct {
 	HttpKeepAlive bool
 	MaxRetries    int
 	HttpProxy     string
+	UseSSL        bool
 }
 
 // Client will returns a client with connections for all product
@@ -66,7 +68,7 @@ func (c *Config) Client() (*KsyunClient, error) {
 	}
 	client.config = c
 	url := &utils.UrlInfo{
-		UseSSL:                      false,
+		UseSSL:                      c.UseSSL,
 		Locate:                      false,
 		CustomerDomain:              c.Domain,
 		CustomerDomainIgnoreService: c.IgnoreService,
@@ -94,6 +96,7 @@ func (c *Config) Client() (*KsyunClient, error) {
 	client.kceconn = kce.SdkNew(cli, cfg, url)
 	client.kcev2conn = kcev2.SdkNew(cli, cfg, url)
 	client.knadconn = knad.SdkNew(cli, cfg, url)
+	client.pdnsconn = pdns.SdkNew(cli, cfg, url)
 
 	// 懒加载ks3-client 所以不在此初始化
 	return &client, nil
@@ -135,6 +138,11 @@ func registerClient(cli *session.Session, c *Config) {
 	cli.Config.Retryer = network.GetKsyunRetryer(c.MaxRetries)
 
 	cli.Handlers.CompleteAttempt.PushBackNamed(network.NetErrorHandler)
+
+	// TODO: output request's information when it encounters the special error that reset connection.
+	// cli.Handlers.CompleteAttempt.PushBackNamed(network.OutputResetError)
+
+	cli.Handlers.Sign.PushBackNamed(network.HandleRequestBody)
 }
 
 func getKsyunClient(c *Config) *http.Client {
@@ -151,14 +159,14 @@ func getKsyunClient(c *Config) *http.Client {
 		},
 		DialContext: (&net.Dialer{
 			Timeout:   30 * time.Second, // dial connect timeout
-			KeepAlive: 5 * time.Second,  // the interval probes time between the ends of network
+			KeepAlive: 30 * time.Second, // the interval probes time between the ends of network
 		}).DialContext,
 		ForceAttemptHTTP2:     false,
-		MaxIdleConns:          30,
-		IdleConnTimeout:       60 * time.Second,
+		MaxIdleConns:          100,
+		IdleConnTimeout:       90 * time.Second,
 		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 3 * time.Second,
-		DisableKeepAlives:     c.HttpKeepAlive,
+		ExpectContinueTimeout: 1 * time.Second,
+		DisableKeepAlives:     !c.HttpKeepAlive,
 	}
 
 	httpClient := &http.Client{